[fixed] #5: "error accessing factory from ..."
During the load of shared object (.so) files, the message "error accessing factory from..." occured and the plug-in failed to load.
I assume that the gcc compiler behaviour changed a while ago, so that a comparison of (NULL == 0) fails. The code in question was in DLLFactoryBase: "if (LastError() == 0 /* is NULL, not 0 */)".

[added] debug information